+++ crates/outlayer-app/core/Cargo.toml
@@ -8,7 +8,6 @@ rust-version.workspace = true
 repository.workspace = true
 
 [dependencies]
-defuse-crypto = { workspace = true, default-features = false, features = ["ed25519"] }
 near-account-id.workspace = true
 
 arbitrary = { workspace = true, features = ["derive"], optional = true }
@@ -17,34 +16,16 @@ cfg_eval = { workspace = true, optional = true }
 defuse-digest = { workspace = true, features = ["sha2"], optional = true }
 schemars = { workspace = true, features = ["derive"], optional = true }
 serde = { workspace = true, features = ["derive"], optional = true }
-serde_with = { workspace = true, features = ["base64", "hex"], optional = true }
+serde_with = { workspace = true, features = ["hex"], optional = true }
 
 [features]
 abi = ["borsh-schema", "schemars-v0_8"]
-arbitrary = [
-  "defuse-crypto/arbitrary",
-  "dep:arbitrary",
-  "near-account-id/arbitrary",
-]
-borsh = ["defuse-crypto/borsh", "dep:borsh", "near-account-id/borsh"]
-borsh-schema = [
-  "borsh",
-  "borsh/unstable__schema",
-  "defuse-crypto/abi",
-  "near-account-id/abi",
-]
+arbitrary = ["dep:arbitrary", "near-account-id/arbitrary"]
+borsh = ["dep:borsh", "near-account-id/borsh"]
+borsh-schema = ["borsh", "borsh/unstable__schema", "near-account-id/abi"]
 digest = ["dep:defuse-digest"]
-parse = ["defuse-crypto/fmt"]
-serde = [
-  "defuse-crypto/serde",
-  "dep:cfg_eval",
-  "dep:serde",
-  "dep:serde_with",
-  "near-account-id/serde",
-  "parse",
-]
+serde = ["dep:cfg_eval", "dep:serde", "dep:serde_with", "near-account-id/serde"]
 schemars-v0_8 = [
-  "defuse-crypto/abi",
   "dep:schemars",
   "near-account-id/schemars-v0_8",
   "serde",
